requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Session
Flask-Caching
Flask-Login
Flask-Mail
Flask-Bcrypt
Authlib
psycopg2-binary
werkzeug
SQLAlchemy
aiohttp
pytest
redis
requests
gunicorn==23.0.0
psycopg2-binary
python-dotenv